# Asset-Manager/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Farm By Gaavwala ecommerce API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface User {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  createdAt: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  /** @nullable */
  phone?: string | null;
  role: string;
  blocked: boolean;
  orderCount?: number;
  totalSpent?: number;
  createdAt: string;
}

export interface AdminUserBlockInput {
  blocked: boolean;
}

export interface Category {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  emoji?: string | null;
  productCount?: number;
}

export interface CategoryInput {
  name: string;
  slug: string;
  description?: string;
  emoji?: string;
}

export interface Product {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
  price: number;
  /** @nullable */
  originalPrice?: number | null;
  /** @nullable */
  discount?: number | null;
  stock: number;
  /** @nullable */
  categoryId?: number | null;
  rating: number;
  reviewCount: number;
  images?: string[];
  /** @nullable */
  badge?: string | null;
  /** @nullable */
  emoji?: string | null;
  featured: boolean;
  active: boolean;
  createdAt?: string;
}

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  originalPrice?: number;
  stock: number;
  categoryId?: number;
  images?: string[];
  badge?: string;
  emoji?: string;
  featured?: boolean;
  active?: boolean;
}

export interface ProductUpdate {
  name?: string;
  slug?: string;
  description?: string;
  price?: number;
  originalPrice?: number;
  stock?: number;
  categoryId?: number;
  images?: string[];
  badge?: string;
  emoji?: string;
  featured?: boolean;
  active?: boolean;
}

export interface ProductsResponse {
  products: Product[];
  total: number;
  page: number;
  limit: number;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface RegisterInput {
  email: string;
  /** @minLength 6 */
  password: string;
  name: string;
  phone?: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface CartItemDetails {
  id: number;
  cartId: number;
  productId: number;
  quantity: number;
  product?: Product | null;
}

export interface Cart {
  id: number;
  userId: number;
  total: number;
  items: CartItemDetails[];
}

export interface CartItemInput {
  productId: number;
  quantity: number;
}

export interface CartItemUpdate {
  quantity: number;
}

export interface ShippingAddress {
  name: string;
  phone: string;
  line1: string;
  /** @nullable */
  line2?: string | null;
  city: string;
  state: string;
  pincode: string;
}

export interface OrderItem {
  id: number;
  orderId: number;
  productId: number;
  quantity: number;
  price: number;
  name: string;
  /** @nullable */
  emoji?: string | null;
}

export interface Order {
  id: number;
  userId: number;
  status: string;
  subtotal: number;
  total: number;
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  paymentStatus: string;
  /** @nullable */
  razorpayOrderId?: string | null;
  /** @nullable */
  razorpayPaymentId?: string | null;
  /** @nullable */
  notes?: string | null;
  items: OrderItem[];
  createdAt: string;
}

export interface OrderInput {
  shippingAddress: ShippingAddress;
  paymentMethod: string;
  notes?: string;
}

export interface WishlistItem {
  id: number;
  userId: number;
  productId: number;
  product?: Product | null;
  createdAt: string;
}

export interface WishlistInput {
  productId: number;
}

export interface RazorpayOrderInput {
  orderId: number;
}

export interface RazorpayOrderResponse {
  razorpayOrderId: string;
  amount: number;
  currency: string;
  keyId: string;
}

export interface RazorpayVerifyInput {
  razorpayOrderId: string;
  razorpayPaymentId: string;
  razorpaySignature: string;
  orderId: number;
}

export interface PaymentVerifyResponse {
  success: boolean;
  message: string;
}

export interface AdminDashboard {
  totalOrders: number;
  totalRevenue: number;
  totalUsers: number;
  totalProducts: number;
  pendingOrders?: number;
  lowStockCount?: number;
  recentOrders?: Order[];
}

export interface AdminOrderUpdate {
  status?: string;
  paymentStatus?: string;
}

export interface MonthlyStats {
  month: string;
  revenue: number;
  orders: number;
}

export type ListProductsParams = {
category?: string;
featured?: boolean;
search?: string;
page?: number;
limit?: number;
};