# Asset-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Farm By Gaavwala ecommerce API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/dashboard:
    get:
      operationId: adminGetDashboard
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDashboard"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/products:
    get:
      operationId: adminListProducts
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
    post:
      operationId: adminCreateProduct
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/products/{id}:
    put:
      operationId: adminUpdateProduct
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
    delete:
      operationId: adminDeleteProduct
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/orders:
    get:
      operationId: adminListOrders
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/orders/{id}:
    put:
      operationId: adminUpdateOrder
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminOrderUpdate"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Not found
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/users/{id}/block:
    patch:
      operationId: adminToggleUserBlock
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminUserBlockInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
  /admin/categories:
    get:
      operationId: adminListCategories
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
    post:
      operationId: adminCreateCategory
      tags: [admin]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
  /admin/categories/{id}:
    put:
      operationId: adminUpdateCategory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
        "401":
          description: Unauthorized
    delete:
      operationId: adminDeleteCategory
      tags: [admin]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
  /admin/stats/monthly:
    get:
      operationId: adminGetMonthlyStats
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MonthlyStats"
        "401":
          description: Unauthorized
  /admin/inventory:
    get:
      operationId: adminGetInventory
      tags: [admin]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
        "401":
          description: Unauthorized
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        createdAt:
          type: string
          format: date-time
    AdminUser:
      type: object
      required: [id, email, name, role, blocked, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
        blocked:
          type: boolean
        orderCount:
          type: integer
        totalSpent:
          type: integer
        createdAt:
          type: string
          format: date-time
    AdminUserBlockInput:
      type: object
      required: [blocked]
      properties:
        blocked:
          type: boolean
    Category:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        emoji:
          type: ["string", "null"]
        productCount:
          type: integer
    CategoryInput:
      type: object
      required: [name, slug]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        emoji:
          type: string
    Product:
      type: object
      required: [id, name, slug, price, stock, rating, reviewCount, featured, active]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: integer
        originalPrice:
          type: ["integer", "null"]
        discount:
          type: ["integer", "null"]
        stock:
          type: integer
        categoryId:
          type: ["integer", "null"]
        rating:
          type: number
        reviewCount:
          type: integer
        images:
          type: array
          items:
            type: string
        badge:
          type: ["string", "null"]
        emoji:
          type: ["string", "null"]
        featured:
          type: boolean
        active:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ProductInput:
      type: object
      required: [name, slug, price, stock]
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: integer
        originalPrice:
          type: integer
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        badge:
          type: string
        emoji:
          type: string
        featured:
          type: boolean
        active:
          type: boolean
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        price:
          type: integer
        originalPrice:
          type: integer
        stock:
          type: integer
        categoryId:
          type: integer
        images:
          type: array
          items:
            type: string
        badge:
          type: string
        emoji:
          type: string
        featured:
          type: boolean
        active:
          type: boolean
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    RegisterInput:
      type: object
      required: [email, password, name]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 6
        name:
          type: string
        phone:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
...
    ShippingAddress:
      type: object
      required: [name, phone, line1, city, state, pincode]
      properties:
        name:
          type: string
        phone:
          type: string
        line1:
          type: string
        line2:
          type: ["string", "null"]
        city:
          type: string
        state:
          type: string
        pincode:
          type: string
    OrderItem:
      type: object
      required: [id, orderId, productId, quantity, price, name]
      properties:
        id:
          type: integer
        orderId:
          type: integer
        productId:
          type: integer
        quantity:
          type: integer
        price:
          type: integer
        name:
          type: string
        emoji:
          type: ["string", "null"]
    Order:
      type: object
      required: [id, userId, status, subtotal, total, shippingAddress, paymentMethod, paymentStatus, items, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        status:
          type: string
        subtotal:
          type: integer
        total:
          type: integer
        shippingAddress:
          $ref: "#/components/schemas/ShippingAddress"
        paymentMethod:
          type: string
        paymentStatus:
          type: string
        razorpayOrderId:
          type: ["string", "null"]
        razorpayPaymentId:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
          format: date-time
...
    AdminDashboard:
      type: object
      required: [totalOrders, totalRevenue, totalUsers, totalProducts]
      properties:
        totalOrders:
          type: integer
        totalRevenue:
          type: integer
        totalUsers:
          type: integer
        totalProducts:
          type: integer
        pendingOrders:
          type: integer
        lowStockCount:
          type: integer
        recentOrders:
          type: array
          items:
            $ref: "#/components/schemas/Order"
    AdminOrderUpdate:
      type: object
      properties:
        status:
          type: string
        paymentStatus:
          type: string
    MonthlyStats:
      type: object
      required: [month, revenue, orders]
      properties:
        month:
          type: string
        revenue:
          type: integer
        orders:
          type: integer