# Asset-Manager/lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Farm By Gaavwala ecommerce API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  name: string;
  slug: string;
  description?: string;
  price: number;
  originalPrice?: number;
  stock: number;
  categoryId?: number;
  images?: string[];
  badge?: string;
  emoji?: string;
  featured?: boolean;
  active?: boolean;
}